Renamed UI state and move it to another folder
