+++ app/src/main/java/rick_and_morty/ui/characters/CharactersUiState.kt
@@ -1,9 +1,9 @@
-package rick_and_morty.data.event
+package rick_and_morty.ui.characters
 
 import rick_and_morty.data.model.CharacterResultsDto
 
-data class UIState (
-    val isLoading : Boolean = false,
+data class CharactersUiState (
+    val isLoading : Boolean = true,
     val characterResults: List<CharacterResultsDto> = emptyList(),
     val failure: Exception? = null
 )
