+++ app/src/main/java/rick_and_morty/ui/characters/CharactersViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
-import rick_and_morty.data.event.UIState
 import rick_and_morty.data.repository.CharacterRepository
 import javax.inject.Inject
 
@@ -16,8 +15,8 @@ class CharactersViewModel @Inject constructor(
 
     private var page = 1
 
-    private val _characters = MutableStateFlow(UIState())
-    val characters: StateFlow<UIState> = _characters
+    private val _characters = MutableStateFlow(CharactersUiState())
+    val characters: StateFlow<CharactersUiState> = _characters
 
     init {
         getCharacters()
